# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for print tooling orders
 * OpenAPI spec version: 0.2.0
 */
export interface LoginRequest {
  password: string;
}

export type LoginResponseTokenType = typeof LoginResponseTokenType[keyof typeof LoginResponseTokenType];


export const LoginResponseTokenType = {
  Bearer: 'Bearer',
} as const;

export interface LoginResponse {
  access_token: string;
  token_type: LoginResponseTokenType;
  expires_in: number;
}

export interface HealthStatus {
  status: string;
}

export type FieldDefinitionType = typeof FieldDefinitionType[keyof typeof FieldDefinitionType];


export const FieldDefinitionType = {
  text: 'text',
  file: 'file',
  number: 'number',
  select: 'select',
  textarea: 'textarea',
} as const;

export interface FieldDefinition {
  key: string;
  label: string;
  type: FieldDefinitionType;
  required: boolean;
  options?: string[];
  readOnly?: boolean;
}

export interface ProductDefinition {
  key: string;
  name: string;
  fields: FieldDefinition[];
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  Получен: 'Получен',
  Ожидает_согласования: 'Ожидает согласования',
  В_производстве: 'В производстве',
  Доставка: 'Доставка',
  Готов_к_отгрузке: 'Готов к отгрузке',
} as const;

export interface OrderSummary {
  id: number;
  order_number: string;
  product_type: string;
  product_name: string;
  client: string;
  contact: string;
  comment: string;
  status: OrderStatus;
  created_at: string;
}

export type OrderData = {[key: string]: string | number | boolean | null | string[]};

export type Order = OrderSummary & {
  data: OrderData;
  files: string[];
};

export interface ErrorResponse {
  error: string;
}

export type CreateOrderBody = {
  product_type: string;
  client?: string;
  contact?: string;
  comment?: string;
  /** JSON object with product-specific values */
  data: string;
  /**
     * Files are uploaded as repeated multipart fields named files. Each file is limited to 45 MiB and the complete request to 50 MiB.
     * @maxItems 10
     */
  files?: Blob[];
};