# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: API for print tooling orders
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Print tooling catalog
  - name: orders
    description: Customer tooling orders
  - name: auth
    description: Operator authentication
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List tooling product types and fields
      responses:
        "200":
          description: Product catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductDefinition"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in as an operator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Operator access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders:
    get:
      operationId: getOrders
      tags: [orders]
      summary: List customer orders
      security:
        - bearerAuth: []
      responses:
        "401":
          description: Missing or invalid operator token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderSummary"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a tooling order
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                product_type:
                  type: string
                client:
                  type: string
                contact:
                  type: string
                comment:
                  type: string
                data:
                  type: string
                  description: JSON object with product-specific values
                files:
                  type: array
                  maxItems: 10
                  description: Files are uploaded as repeated multipart fields named files. Each file is limited to 45 MiB and the complete request to 50 MiB.
                  items:
                    type: string
                    format: binary
              required:
                - product_type
                - data
      responses:
        "401":
          description: Missing or invalid operator token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          description: Invalid order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "413":
          description: Uploaded files exceed the 50 MiB request limit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many order attempts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /orders/{order_id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "401":
          description: Missing or invalid operator token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /uploads/{filename}:
    get:
      operationId: getUpload
      tags: [orders]
      summary: Download an uploaded order file
      security:
        - bearerAuth: []
      parameters:
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "401":
          description: Missing or invalid operator token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: Uploaded file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OperatorAccessToken
  schemas:
    LoginRequest:
      type: object
      properties:
        password:
          type: string
          format: password
      required:
        - password
    LoginResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
            type: number
      required:
        - access_token
        - token_type
        - expires_in
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    FieldDefinition:
      type: object
      properties:
        key:
          type: string
        label:
          type: string
        type:
          type: string
          enum: [text, file, number, select, textarea]
        required:
          type: boolean
        options:
          type: array
          items:
            type: string
        readOnly:
          type: boolean
      required:
        - key
        - label
        - type
        - required
    ProductDefinition:
      type: object
      properties:
        key:
          type: string
        name:
          type: string
        fields:
          type: array
          items:
            $ref: "#/components/schemas/FieldDefinition"
      required:
        - key
        - name
        - fields
    OrderSummary:
      type: object
      properties:
        id:
          type: number
        order_number:
          type: string
        product_type:
          type: string
        product_name:
          type: string
        client:
          type: string
        contact:
          type: string
        comment:
          type: string
        status:
          $ref: "#/components/schemas/OrderStatus"
        created_at:
          type: string
      required:
        - id
        - order_number
        - product_type
        - product_name
        - client
        - contact
        - comment
        - status
        - created_at
    OrderStatus:
      type: string
      enum:
        - Получен
        - Ожидает согласования
        - В производстве
        - Доставка
        - Готов к отгрузке
    Order:
      allOf:
        - $ref: "#/components/schemas/OrderSummary"
        - type: object
          properties:
            data:
              type: object
              additionalProperties:
                description: Product-specific values plus the __file_fields string array.
                anyOf:
                  - type: string
                  - type: number
                  - type: boolean
                  - type: "null"
                  - type: array
                    items:
                      type: string
            files:
              type: array
              items:
                type: string
          required:
            - data
            - files
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error