# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API for print tooling orders
 * OpenAPI spec version: 0.2.0
 */
import type { LoginResponseTokenType } from './loginResponseTokenType';

export interface LoginResponse {
  access_token: string;
  token_type: LoginResponseTokenType;
  expires_in: number;
}